stroylux/requirements.txt
amqp==2.5.2
aniso8601==7.0.0
asgiref==3.2.7
Babel==2.7.0
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
dj-email-url==1.0.0
Django==3.0.5
django-appconf==1.0.4
django-countries==5.5
django-debug-toolbar==2.2
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-js-asset==1.2.2
django-measurement==3.2.0
django-mptt==0.11.0
django-phonenumber-field==2.4.0
django-prices==2.2.0
django-recaptcha==2.0.6
django-render-block==0.6
django-storages==1.9.1
django-templated-email==2.3.0
django-versatileimagefield==2.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
et-xmlfile==1.0.1
google-measurement-protocol==1.0.0
graphene==2.1.8
graphene-django==2.9.0
graphene-django-optimizer==0.6.2
graphene-federation==0.0.3
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.0
kombu==4.6.8
maxminddb==1.5.2
maxminddb-geolite2==2018.703
measurement==3.2.0
mpmath==1.1.0
oauthlib==3.1.0
openpyxl==3.0.3
opentracing==2.3.0
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==7.1.1
prices==1.0.0
promise==2.3
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-magic==0.4.15;sys_platform == 'darwin'
python-magic-bin;sys_platform == 'win32'
pytz==2019.3
redis==3.4.1
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
sympy==1.5.1
text-unidecode==1.3
traitlets==4.3.3
transliterate==1.10.2
typing==3.7.4.1
urllib3==1.25.9
uWSGI==2.0.18;sys_platform == 'darwin'
vine==1.3.0
wcwidth==0.1.9
